New state STATE_IGNORED for syncPathState()
Now syncPathState returns MegaApi::STATE_NONE, MegaApi::STATE_SYNCED,
MegaApi::STATE_PENDING, MegaApi::STATE_SYNCING or
MegaApi::STATE_IGNORED.
MegaApi::STATE_NONE means that the path is out of a synced folder.
MegaApi::STATE_IGNORED means that the path is inside a synced folder but is excluded by the synchronization algorithm.
Other states preserve their previous meanings.

+++ src/megaapi.cpp
@@ -565,7 +565,7 @@ bool MegaApi::moveToLocalDebris(const char *path)
 
 int MegaApi::syncPathState(string* path)
 {
-    return (int)pImpl->syncPathState(path);
+    return pImpl->syncPathState(path);
 }
 
 MegaNode *MegaApi::getSyncedNode(string *path)
